Move accessValidations to common dir for routes
